Converted x86 to uint16_t

+++ x86/lwe_original.h
@@ -14,7 +14,8 @@
 
 void fwd_ntt(int a_0[], int a_1[]);
 void a_gen( int a_0[],  int a_1[]);
-int compare2(int a_0[128],int a_1[128],int large[M]);
+//int compare2(int a_0[128],int a_1[128],int large[M]);
+int compare2(int a_0[128],int a_1[128],uint16_t large[M]);
 
 
 #endif /* LWE_ORIGINAL_H_ */
